Correcting test case failure
